rename context provider and remove some unnecessary comments
