+++ src/theme/Root.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {SearchPopoverProvider} from '../components/SearchProvider';
+import {SearchPopoverProvider} from '../components/SearchPopoverProvider';
 
 // Default implementation, that you can customize
 export default function Root({children}) {
